replace textResponsesRedux with storing in component state
