+++ apps/src/sites/studio/pages/teacher_dashboard/show.js
@@ -21,7 +21,6 @@ import teacherSections, {
 } from '@cdo/apps/templates/teacherDashboard/teacherSectionsRedux';
 import sectionData, {setSection} from '@cdo/apps/redux/sectionDataRedux';
 import stats from '@cdo/apps/templates/teacherDashboard/statsRedux';
-import textResponses from '@cdo/apps/templates/textResponses/textResponsesRedux';
 import sectionAssessments from '@cdo/apps/templates/sectionAssessments/sectionAssessmentsRedux';
 import sectionProgress from '@cdo/apps/templates/sectionProgress/sectionProgressRedux';
 import sectionStandardsProgress from '@cdo/apps/templates/sectionProgress/standards/sectionStandardsProgressRedux';
@@ -57,7 +56,6 @@ $(document).ready(function() {
     sectionProgress,
     unitSelection,
     stats,
-    textResponses,
     sectionAssessments,
     currentUser,
     sectionStandardsProgress,
